publish script

+++ my-scripts/build-env/codeblocks/removeDir.ts
@@ -1,4 +1,4 @@
-import { existsSync, lstatSync, mkdirSync, rmdir as rmdirAsync, Stats, unlink as unlinkAsync } from 'fs';
+import { rmdir as rmdirAsync, unlink as unlinkAsync } from 'fs';
 import * as rimraf from 'rimraf';
 import { isWin } from '../misc/constants';
 import { globalSuccessMessage } from '../misc/globalOutput';
@@ -16,11 +16,17 @@ function wrapFs(of: Function, output: NodeJS.WritableStream): Function {
 
 export function removeDirectory(path: string, output: NodeJS.WritableStream, verbose = true) {
 	output.write(`removing directory: ${path}...\n`);
+	
+	if (process.cwd().indexOf(path) === 0) {
+		output.write(`  Cwd: ${process.cwd()}\n`);
+		return Promise.reject(new Error('No way to remove current directory.'));
+	}
+	
 	let p = new Promise<void>((resolve, reject) => {
 		const wrappedCallback = (err) => err? reject(err) : resolve();
 		
 		rimraf(path, {
-			maxBusyTries: 5,
+			maxBusyTries: 20,
 			emfileWait: true,
 			disableGlob: true,
 			unlink: wrapFs(unlinkAsync, output) as typeof unlinkAsync,
